+++ skydentity/policies/checker/policy_checker.py
@@ -1,0 +1,4 @@
+class PolicyChecker:
+    """
+    A policy checker is responsible for appropriately checking a policy.
+    """
